+++ CHANGELOG.md
@@ -11,7 +11,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added generations of GitHub actions
 - Switch for `overleaf`. If set, `_latexmkrc` is generated instead of `latexmkrc`. [lncs#40](https://github.com/latextemplates/LNCS/issues/40)
-- Correct `\keywords{}` syntax for LCNS. [lncs#39](https://github.com/latextemplates/LNCS/issues/39)
 
 ### Changed
 
@@ -21,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed reference to examples in IEEE template
 - Fixed run for TeXLive 2021
+- Correct `\keywords{}` syntax for LCNS. [lncs#39](https://github.com/latextemplates/LNCS/issues/39)
 
 ## [0.5.0] – 2022-02-28